Add basic v1/v2/v3 loading support and simple tests
