+++ cli/compose/loader/testdata/example1.env
@@ -4,5 +4,5 @@ FOO=foo_from_env_file
 # overridden in example2.env
 BAR=bar_from_env_file
 
-# overridden in full-example.yml
+# overridden in v3-full-example.yml
 BAZ=baz_from_env_file
