+++ cli/compose/loader/testdata/v2-full.yml
@@ -1,0 +1,23 @@
+version: "2"
+
+volumes:
+  data:
+    driver: local
+
+networks:
+  front: {}
+
+services:
+  web:
+    build: .
+    networks:
+      - front
+      - default
+    volumes_from:
+      - other
+
+  other:
+    image: busybox:latest
+    command: top
+    volumes:
+      - /data
